Debin files fixed & debug mode for makeDeb.sh

+++ distr/deb/makeDeb.sh
@@ -14,6 +14,7 @@ Options
   --ver=VERSION         razor version
   -S|--sign             sign a result files
   -s|--source           build a source package, if ommited build a binary package
+  --debug               debug mode not build package only create debian directory
 HELP_TEXT
 }
 
@@ -21,9 +22,16 @@ HELP_TEXT
 
 VARIABLES=""
 
+function debug
+{
+    [ -n "$DEBUG" ] && echo $@ >&2
+}
+
 function setVariable
 {
     VARIABLES="${VARIABLES}$2=$3\n"
+    debug "set variable: $2 = $3"
+    debug "variables $VARIABLES"
 }
 
 function getVariable
@@ -218,6 +226,11 @@ while [ $# -gt 0 ]; do
         shift
       ;;
 
+    --debug)
+        DEBUG=1
+        shift
+      ;;
+
     --)
         shift
         break
@@ -283,11 +296,16 @@ if [ -z "$DEBEMAIL" ]; then
   DEBEMAIL=${EMAIL}
 fi
 
+if [ -n "$DEBUG" ]; then
+  TYPE="Debug"
+fi
+
 echo "*******************************"
 echo " Name: ${NAME}"
 echo " Ver:  ${VER}"
-[ "${TYPE}" = "-b" ] && echo " Type: binary"
-[ "${TYPE}" = "-S" ] && echo " Type: source"
+[ "${TYPE}" = "-b"    ] && echo " Type: binary"
+[ "${TYPE}" = "-S"    ] && echo " Type: source"
+[ "${TYPE}" = "Debug" ] && echo " Type: debug"
 echo " Distrib: ${DISTRIB}"
 echo " Release: ${RELEASE}"
 echo " Src dir: ${SRC_DIR}"
@@ -300,11 +318,16 @@ mkdir -p ${OUT_DIR} || exit 2
 DIR=${OUT_DIR}/${NAME}-${VER}
 rm -rf ${DIR}
 
-cp -r ${SRC_DIR} ${DIR}
-rm -rf ${DIR}/.git \
-       ${DIR}/build
+if [ -z "$DEBUG" ]; then
+  cp -r ${SRC_DIR} ${DIR}
+  rm -rf ${DIR}/.git \
+         ${DIR}/build
 
-cd ${DIR}/.. && tar cjf ${NAME}_${VER}.orig.tar.bz2 ${NAME}-${VER}
+  cd ${DIR}/.. && tar cjf ${NAME}_${VER}.orig.tar.bz2 ${NAME}-${VER}
+else
+  mkdir -p ${DIR}/distr
+  cp -r ${SRC_DIR}/distr/ ${DIR}
+fi
 
 for RELEASE in ${RELEASE}; do
     # Debin directory .....................
@@ -321,9 +344,14 @@ for RELEASE in ${RELEASE}; do
         chmod --reference "${src}" ${dest}
     done
     # Debin directory .....................
-    cd ${DIR} && debuild ${TYPE} ${SIGN} -rfakeroot
-    ret=$?
-    [ $ret -eq 0 ] || exit $ret
+
+    if [ -z "$DEBUG" ]; then
+      cd ${DIR} && debuild ${TYPE} ${SIGN} -rfakeroot
+      ret=$?
+      [ $ret -eq 0 ] || exit $ret
+    else
+      exit
+    fi
 done
 
 if [ "${TYPE}" = '-b' ]; then
